auth-example/requirements.txt
Flask==1.1.1
requests==2.22.0
flask_bootstrap==3.3.7.1
flask-login==0.5.0
google-cloud-logging==1.15.0